moved canvas into Waveform folder and added data object
