+++ src/components/Waveform/WaveformData.js
@@ -1,0 +1,43 @@
+import React from 'react';
+
+/*
+ * This is the database of data points for specific waveforms 
+ *
+ * duration: duration of the snippet in ms
+ * nPoints: number of points in the duration
+ * range: the minimum and maximum y values of the given waveform snippet
+ * dataPoints: array of the data points for the snippet
+ *
+ */
+export var waveformData = {
+  'NSR': {
+    'duration': 800,
+    'nPoints': 300,
+    'range': {
+      'min': -0.535,
+      'max': 0.94
+    },
+    'dataPoints' : [-0.275, -0.27, -0.25, -0.25, -0.255, -0.225, -0.22, -0.205, -0.2, -0.205, -0.215, -0.23, -0.22, -0.225, -0.225, -0.225, -0.23, -0.235, -0.24, -0.235, -0.22, -0.21, -0.205, -0.245, -0.285, -0.285, -0.3, -0.31, -0.33, -0.33, -0.325, -0.315, -0.32, -0.315, -0.325, -0.34, -0.345, -0.34, -0.34, -0.35, -0.345, -0.355, -0.33, -0.335, -0.33, -0.32, -0.345, -0.355, -0.34, -0.33, -0.325, -0.33, -0.35, -0.365, -0.36, -0.38, -0.425, -0.445, -0.475, -0.51, -0.535, -0.505, -0.415, -0.3, -0.16, -0.015, 0.235, 0.49, 0.72, 0.875, 0.94, 0.905, 0.755, 0.49, 0.165, -0.11, -0.27, -0.39, -0.45, -0.475, -0.455, -0.425, -0.39, -0.39, -0.385, -0.39, -0.38, -0.38, -0.38, -0.395, -0.385, -0.385, -0.385, -0.375, -0.395, -0.41, -0.41, -0.4, -0.395, -0.39, -0.405, -0.395, -0.385, -0.375, -0.39, -0.39, -0.405, -0.41, -0.41, -0.39, -0.39, -0.395, -0.405, -0.415, -0.4, -0.41, -0.405, -0.41, -0.415, -0.41, -0.4, -0.4, -0.395, -0.39, -0.405, -0.41, -0.39, -0.39, -0.385, -0.385, -0.41, -0.405, -0.395, -0.39, -0.375, -0.39, -0.395, -0.41, -0.4, -0.39, -0.39, -0.385, -0.405, -0.415, -0.415, -0.4, -0.395, -0.405, -0.415, -0.42, -0.42, -0.41, -0.415, -0.425, -0.42, -0.435, -0.43, -0.43, -0.42, -0.43, -0.45, -0.455, -0.45, -0.435, -0.445, -0.45, -0.455, -0.47, -0.46, -0.455, -0.45, -0.455, -0.47, -0.475, -0.46, -0.45, -0.445, -0.44, -0.435, -0.44, -0.41, -0.395, -0.37, -0.365, -0.36, -0.365, -0.34, -0.325, -0.315, -0.32, -0.33, -0.33, -0.32, -0.31, -0.3, -0.3, -0.32, -0.32, -0.315, -0.305, -0.305, -0.295, -0.32, -0.33, -0.305, -0.31, -0.3, -0.3, -0.32, -0.325, -0.31, -0.305, -0.315, -0.305, -0.315, -0.315, -0.31, -0.295, -0.29, -0.305, -0.31, -0.32, -0.315, -0.3, -0.315, -0.315, -0.315, -0.33, -0.315, -0.32, -0.315, -0.325, -0.335, -0.34, -0.335, -0.335, -0.33, -0.325, -0.345, -0.35, -0.345, -0.335, -0.33, -0.33, -0.345, -0.345, -0.345, -0.32, -0.33, -0.335, -0.34, -0.355, -0.335, -0.33, -0.33, -0.335, -0.355, -0.36, -0.355, -0.35, -0.34, -0.345, -0.345, -0.345, -0.345, -0.33, -0.33, -0.335, -0.345, -0.35, -0.35, -0.34, -0.33, -0.345, -0.345, -0.355, -0.35, -0.34, -0.33, -0.34, -0.34, -0.34, -0.33, -0.335, -0.33, -0.335, -0.345, -0.345, -0.34, -0.33, -0.315, -0.295, -0.3, -0.295, -0.285, -0.275, -0.265, -0.265, -0.265, -0.255]
+  },
+  'VT': {
+    'duration': 0,
+    'nPoints': 0,
+    'dataPoints' : []
+  },
+  'VF': {
+    'duration': 0,
+    'nPoints': 0,
+    'dataPoints' : []
+
+  },
+  'PEA': {
+    'duration': 0,
+    'nPoints': 0,
+    'dataPoints' : []
+  },
+  'ASYSTOLE': {
+    'duration': 0,
+    'nPoints': 0,
+    'dataPoints' : []
+  }
+}
